refactor: move shared crypto to @chroxy/store-core with platform adapter
Extract near-identical crypto.ts implementations from app and dashboard
into packages/store-core/src/crypto.ts.

- Add initPRNG(getRandomBytes) for platform-specific PRNG injection
- Add tweetnacl + tweetnacl-util as store-core dependencies
- Export all crypto types and functions from store-core index
- App wrapper calls initPRNG with expo-crypto's getRandomBytes
- Dashboard wrapper re-exports from store-core (browser crypto auto-inits)

Closes #2397
